Add files via upload
The Zomato Exploratory Data Analysis (EDA) project is a shining example of analytical research inside the wide expanse of a dataset derived from the renowned gourmet the centre, Zomato, in the dynamic world of data analysis. We set out on a quest to uncover hidden narratives, interpret latent trends, and understand the mysterious language expressed by information points using the techniques of EDA. The process of interpreting Zomato's facts into an enticing platter of knowledge is captured in this narrative.